backend/requirements.txt
# FastAPI Framework
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Machine Learning
tensorflow==2.16.1
pillow==10.4.0
numpy==1.26.4

# Data Validation
pydantic==2.9.2
pydantic-settings==2.5.2

# Caching
redis==5.0.8

# CORS
fastapi-cors==0.0.6

# Environment
python-dotenv==1.0.1

# Monitoring
prometheus-client==0.21.0

# Testing
pytest==8.3.3
httpx==0.27.2
